Fixes to clear apiIssues status
ApiIssues and error message string.

Tony - for the weather device, zipcode may not be unique if a customer has two structures in same zipcode.  The location name should be unique regardless.   I did not make this change but I think you should consider it for the weather device dni name.